twitter_streaming/requirements.txt
requests
pymongo
python-dotenv
